+++ src/raw/iter.rs
@@ -5,13 +5,13 @@ use crate::{
     oid::ObjectId,
     raw::{Error, ErrorKind, Result},
     spec::{BinarySubtype, ElementType},
-    DateTime, Decimal128, Timestamp,
+    DateTime, Decimal128, RawArray, RawBinaryRef, RawDbPointerRef, RawJavaScriptCodeWithScopeRef,
+    RawRegexRef, Timestamp,
 };
 
 use super::{
-    bson_ref::RawDbPointerRef, checked_add, error::try_with_key, f64_from_slice, i32_from_slice,
-    i64_from_slice, read_lenencoded, read_nullterminated, RawArray, RawBinaryRef, RawBsonRef,
-    RawDocument, RawJavaScriptCodeWithScopeRef, RawRegexRef,
+    checked_add, error::try_with_key, f64_from_slice, i32_from_slice, i64_from_slice,
+    read_len_and_end, read_lenencoded, read_nullterminated, RawBsonRef, RawDocument,
 };
 
 /// An iterator over the document's entries.
@@ -36,56 +36,180 @@ impl<'a> Iter<'a> {
     fn verify_enough_bytes(&self, start: usize, num_bytes: usize) -> Result<()> {
         let end = checked_add(start, num_bytes)?;
         if self.doc.as_bytes().get(start..end).is_none() {
-            return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                message: format!(
-                    "length exceeds remaining length of buffer: {} vs {}",
-                    num_bytes,
-                    self.doc.as_bytes().len() - start
-                ),
-            }));
+            return Err(Error::new_without_key(ErrorKind::new_malformed(format!(
+                "length exceeds remaining length of buffer: {} vs {}",
+                num_bytes,
+                self.doc.as_bytes().len() - start
+            ))));
         }
         Ok(())
     }
 
-    fn next_oid(&self, starting_at: usize) -> Result<ObjectId> {
-        self.verify_enough_bytes(starting_at, 12)?;
-        let oid = ObjectId::from_bytes(
-            self.doc.as_bytes()[starting_at..(starting_at + 12)]
-                .try_into()
-                .unwrap(), // ok because we know slice is 12 bytes long
-        );
-        Ok(oid)
-    }
-
-    fn next_document(&self, starting_at: usize) -> Result<&'a RawDocument> {
+    fn next_document_len(&self, starting_at: usize) -> Result<usize> {
         self.verify_enough_bytes(starting_at, MIN_BSON_DOCUMENT_SIZE as usize)?;
         let size = i32_from_slice(&self.doc.as_bytes()[starting_at..])? as usize;
 
         if size < MIN_BSON_DOCUMENT_SIZE as usize {
-            return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                message: format!("document too small: {} bytes", size),
-            }));
+            return Err(Error::new_without_key(ErrorKind::new_malformed(format!(
+                "document too small: {} bytes",
+                size
+            ))));
         }
 
         self.verify_enough_bytes(starting_at, size)?;
-        let end = starting_at + size;
 
-        if self.doc.as_bytes()[end - 1] != 0 {
-            return Err(Error {
-                key: None,
-                kind: ErrorKind::MalformedValue {
-                    message: "not null terminated".into(),
-                },
-            });
+        if self.doc.as_bytes()[starting_at + size - 1] != 0 {
+            return Err(Error::new_without_key(ErrorKind::new_malformed(
+                "not null terminated",
+            )));
         }
-        RawDocument::from_bytes(&self.doc.as_bytes()[starting_at..end])
+        Ok(size)
+    }
+}
+
+#[derive(Clone)]
+pub struct RawLazyElement<'a> {
+    pub key: &'a str,
+    pub kind: ElementType,
+    doc: &'a RawDocument,
+    start_at: usize,
+    size: usize,
+}
+
+impl<'a> RawLazyElement<'a> {
+    pub fn resolve(self) -> Result<RawBsonRef<'a>> {
+        Ok(match self.kind {
+            ElementType::Null => RawBsonRef::Null,
+            ElementType::Undefined => RawBsonRef::Undefined,
+            ElementType::MinKey => RawBsonRef::MinKey,
+            ElementType::MaxKey => RawBsonRef::MaxKey,
+            ElementType::ObjectId => RawBsonRef::ObjectId(ObjectId::from_bytes(
+                self.doc.as_bytes()[self.start_at..(self.start_at + self.size)]
+                    .try_into()
+                    .map_err(|e| Error::new_with_key(self.key, ErrorKind::new_malformed(e)))?,
+            )),
+            ElementType::EmbeddedDocument => RawBsonRef::Document(RawDocument::from_bytes(
+                &self.doc.as_bytes()[self.start_at..(self.start_at + self.size)],
+            )?),
+            ElementType::Array => RawBsonRef::Array(RawArray::from_doc(RawDocument::from_bytes(
+                &self.doc.as_bytes()[self.start_at..(self.start_at + self.size)],
+            )?)),
+            ElementType::Int32 => {
+                RawBsonRef::Int32(i32_from_slice(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::Int64 => {
+                RawBsonRef::Int64(i64_from_slice(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::Double => {
+                RawBsonRef::Double(f64_from_slice(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::String => {
+                RawBsonRef::String(read_lenencoded(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::Boolean => RawBsonRef::Boolean(
+                read_bool(&self.doc.as_bytes()[self.start_at..])
+                    .map_err(|e| Error::new_with_key(self.key, ErrorKind::new_malformed(e)))?,
+            ),
+            ElementType::DateTime => RawBsonRef::DateTime(DateTime::from_millis(i64_from_slice(
+                &self.doc.as_bytes()[self.start_at..],
+            )?)),
+            ElementType::Decimal128 => RawBsonRef::Decimal128(Decimal128::from_bytes(
+                self.doc.as_bytes()[self.start_at..(self.start_at + self.size)]
+                    .try_into()
+                    .map_err(|e| Error::new_with_key(self.key, ErrorKind::new_malformed(e)))?,
+            )),
+            ElementType::JavaScriptCode => {
+                RawBsonRef::JavaScriptCode(read_lenencoded(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::Symbol => {
+                RawBsonRef::Symbol(read_lenencoded(&self.doc.as_bytes()[self.start_at..])?)
+            }
+            ElementType::DbPointer => RawBsonRef::DbPointer(RawDbPointerRef {
+                namespace: read_lenencoded(&self.doc.as_bytes()[self.start_at..])?,
+                id: ObjectId::from_bytes(
+                    self.doc.as_bytes()[self.start_at..(self.start_at + self.size)]
+                        .try_into()
+                        .map_err(|e| Error::new_with_key(self.key, ErrorKind::new_malformed(e)))?,
+                ),
+            }),
+            ElementType::RegularExpression => {
+                let pattern = read_nullterminated(&self.doc.as_bytes()[self.start_at..])?;
+                let options = read_nullterminated(
+                    &self.doc.as_bytes()[(self.start_at + pattern.len() + 1)..],
+                )?;
+                RawBsonRef::RegularExpression(RawRegexRef { pattern, options })
+            }
+            ElementType::Timestamp => RawBsonRef::Timestamp(
+                Timestamp::from_reader(&self.doc.as_bytes()[self.start_at..])
+                    .map_err(|e| Error::new_with_key(self.key, ErrorKind::new_malformed(e)))?,
+            ),
+            ElementType::Binary => {
+                let len = i32_from_slice(&self.doc.as_bytes()[self.start_at..])? as usize;
+                let data_start = self.start_at + 4 + 1;
+
+                if len >= i32::MAX as usize {
+                    return Err(Error::new_with_key(
+                        self.key,
+                        ErrorKind::new_malformed(format!("binary length exceeds maximum: {}", len)),
+                    ));
+                }
+
+                let subtype = BinarySubtype::from(self.doc.as_bytes()[self.start_at + 4]);
+                let data = match subtype {
+                    BinarySubtype::BinaryOld => {
+                        if len < 4 {
+                            return Err(Error::new_with_key(
+                                self.key,
+                                ErrorKind::new_malformed(
+                                    "old binary subtype has no inner declared length",
+                                ),
+                            ));
+                        }
+                        let oldlength =
+                            i32_from_slice(&self.doc.as_bytes()[data_start..])? as usize;
+                        if checked_add(oldlength, 4)? != len {
+                            return Err(Error::new_with_key(
+                                self.key,
+                                ErrorKind::new_malformed(
+                                    "old binary subtype has wrong inner declared length",
+                                ),
+                            ));
+                        }
+                        &self.doc.as_bytes()[(data_start + 4)..(data_start + len)]
+                    }
+                    _ => &self.doc.as_bytes()[data_start..(data_start + len)],
+                };
+                RawBsonRef::Binary(RawBinaryRef {
+                    subtype,
+                    bytes: data,
+                })
+            }
+            ElementType::JavaScriptCodeWithScope => {
+                let slice = &self.doc.as_bytes()[self.start_at..(self.start_at + self.size)];
+                let code = read_lenencoded(&slice[4..])?;
+                let scope_start = 4 + 4 + code.len() + 1;
+                let scope = RawDocument::from_bytes(&slice[scope_start..])?;
+
+                RawBsonRef::JavaScriptCodeWithScope(RawJavaScriptCodeWithScopeRef { code, scope })
+            }
+        })
+    }
+}
+
+impl<'a> Iter<'a> {
+    pub(crate) fn into_eager(self) -> Box<impl Iterator<Item = Result<(&'a str, RawBsonRef<'a>)>>> {
+        Box::new(self.map(|outer| {
+            outer.and_then(|val| -> Result<(&'a str, RawBsonRef<'a>)> {
+                Ok((val.key, val.resolve()?))
+            })
+        }))
     }
 }
 
 impl<'a> Iterator for Iter<'a> {
-    type Item = Result<(&'a str, RawBsonRef<'a>)>;
+    type Item = Result<RawLazyElement<'a>>;
 
-    fn next(&mut self) -> Option<Result<(&'a str, RawBsonRef<'a>)>> {
+    fn next(&mut self) -> Option<Result<RawLazyElement<'a>>> {
         if !self.valid {
             return None;
         } else if self.offset == self.doc.as_bytes().len() - 1 {
@@ -94,18 +218,15 @@ impl<'a> Iterator for Iter<'a> {
                 return None;
             } else {
                 self.valid = false;
-                return Some(Err(Error {
-                    key: None,
-                    kind: ErrorKind::MalformedValue {
-                        message: "document not null terminated".into(),
-                    },
-                }));
+                return Some(Err(Error::new_without_key(ErrorKind::new_malformed(
+                    "document not null terminated",
+                ))));
             }
         } else if self.offset >= self.doc.as_bytes().len() {
             self.valid = false;
-            return Some(Err(Error::new_without_key(ErrorKind::MalformedValue {
-                message: "iteration overflowed document".to_string(),
-            })));
+            return Some(Err(Error::new_without_key(ErrorKind::new_malformed(
+                "iteration overflowed document",
+            ))));
         }
 
         let key = match read_nullterminated(&self.doc.as_bytes()[self.offset + 1..]) {
@@ -116,193 +237,87 @@ impl<'a> Iterator for Iter<'a> {
             }
         };
 
+        let valueoffset = self.offset + 1 + key.len() + 1; // type specifier + key + \0
         let kvp_result = try_with_key(key, || {
-            let valueoffset = self.offset + 1 + key.len() + 1; // type specifier + key + \0
-
             let element_type = match ElementType::from(self.doc.as_bytes()[self.offset]) {
                 Some(et) => et,
                 None => {
                     return Err(Error::new_with_key(
                         key,
-                        ErrorKind::MalformedValue {
-                            message: format!("invalid tag: {}", self.doc.as_bytes()[self.offset]),
-                        },
+                        ErrorKind::new_malformed(format!(
+                            "invalid tag: {}",
+                            self.doc.as_bytes()[self.offset]
+                        )),
                     ))
                 }
             };
 
-            let (element, element_size) = match element_type {
-                ElementType::Int32 => {
-                    let i = i32_from_slice(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::Int32(i), 4)
-                }
-                ElementType::Int64 => {
-                    let i = i64_from_slice(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::Int64(i), 8)
-                }
-                ElementType::Double => {
-                    let f = f64_from_slice(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::Double(f), 8)
-                }
-                ElementType::String => {
-                    let s = read_lenencoded(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::String(s), 4 + s.len() + 1)
-                }
-                ElementType::EmbeddedDocument => {
-                    let doc = self.next_document(valueoffset)?;
-                    (RawBsonRef::Document(doc), doc.as_bytes().len())
-                }
-                ElementType::Array => {
-                    let doc = self.next_document(valueoffset)?;
-                    (
-                        RawBsonRef::Array(RawArray::from_doc(doc)),
-                        doc.as_bytes().len(),
-                    )
-                }
+            let element_size = match element_type {
+                ElementType::Int32 => 4,
+                ElementType::Int64 => 8,
+                ElementType::Double => 8,
+                ElementType::String => read_len_and_end(&self.doc.as_bytes()[valueoffset..])?.0,
+                ElementType::EmbeddedDocument => self.next_document_len(valueoffset)?,
+                ElementType::Array => self.next_document_len(valueoffset)?,
                 ElementType::Binary => {
-                    let len = i32_from_slice(&self.doc.as_bytes()[valueoffset..])? as usize;
-                    let data_start = valueoffset + 4 + 1;
-
-                    if len >= i32::MAX as usize {
-                        return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                            message: format!("binary length exceeds maximum: {}", len),
-                        }));
-                    }
-
-                    self.verify_enough_bytes(valueoffset + 4, len + 1)?;
-                    let subtype = BinarySubtype::from(self.doc.as_bytes()[valueoffset + 4]);
-                    let data = match subtype {
-                        BinarySubtype::BinaryOld => {
-                            if len < 4 {
-                                return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                                    message: "old binary subtype has no inner declared length"
-                                        .into(),
-                                }));
-                            }
-                            let oldlength =
-                                i32_from_slice(&self.doc.as_bytes()[data_start..])? as usize;
-                            if checked_add(oldlength, 4)? != len {
-                                return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                                    message: "old binary subtype has wrong inner declared length"
-                                        .into(),
-                                }));
-                            }
-                            &self.doc.as_bytes()[(data_start + 4)..(data_start + len)]
-                        }
-                        _ => &self.doc.as_bytes()[data_start..(data_start + len)],
-                    };
-                    (
-                        RawBsonRef::Binary(RawBinaryRef {
-                            subtype,
-                            bytes: data,
-                        }),
-                        4 + 1 + len,
-                    )
+                    (i32_from_slice(&self.doc.as_bytes()[valueoffset..])? + 4 + 1) as usize
                 }
-                ElementType::ObjectId => {
-                    let oid = self.next_oid(valueoffset)?;
-                    (RawBsonRef::ObjectId(oid), 12)
-                }
-                ElementType::Boolean => {
-                    let b = read_bool(&self.doc.as_bytes()[valueoffset..]).map_err(|e| {
-                        Error::new_with_key(
-                            key,
-                            ErrorKind::MalformedValue {
-                                message: e.to_string(),
-                            },
-                        )
-                    })?;
-                    (RawBsonRef::Boolean(b), 1)
-                }
-                ElementType::DateTime => {
-                    let ms = i64_from_slice(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::DateTime(DateTime::from_millis(ms)), 8)
-                }
+                ElementType::ObjectId => 12,
+                ElementType::Boolean => 1,
+                ElementType::DateTime => 8,
                 ElementType::RegularExpression => {
                     let pattern = read_nullterminated(&self.doc.as_bytes()[valueoffset..])?;
                     let options = read_nullterminated(
                         &self.doc.as_bytes()[(valueoffset + pattern.len() + 1)..],
                     )?;
-                    (
-                        RawBsonRef::RegularExpression(RawRegexRef { pattern, options }),
-                        pattern.len() + 1 + options.len() + 1,
-                    )
+                    pattern.len() + 1 + options.len() + 1
                 }
-                ElementType::Null => (RawBsonRef::Null, 0),
-                ElementType::Undefined => (RawBsonRef::Undefined, 0),
-                ElementType::Timestamp => {
-                    let ts = Timestamp::from_reader(&self.doc.as_bytes()[valueoffset..]).map_err(
-                        |e| {
-                            Error::new_without_key(ErrorKind::MalformedValue {
-                                message: e.to_string(),
-                            })
-                        },
-                    )?;
-                    (RawBsonRef::Timestamp(ts), 8)
-                }
+                ElementType::Null => 0,
+                ElementType::Undefined => 0,
+                ElementType::Timestamp => 8,
+                ElementType::Symbol => read_len_and_end(&self.doc.as_bytes()[valueoffset..])?.0,
                 ElementType::JavaScriptCode => {
-                    let code = read_lenencoded(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::JavaScriptCode(code), 4 + code.len() + 1)
+                    read_len_and_end(&self.doc.as_bytes()[valueoffset..])?.0
                 }
                 ElementType::JavaScriptCodeWithScope => {
-                    let length = i32_from_slice(&self.doc.as_bytes()[valueoffset..])? as usize;
+                    let length = (i32_from_slice(&self.doc.as_bytes()[valueoffset..])?) as usize;
 
                     if length < MIN_CODE_WITH_SCOPE_SIZE as usize {
-                        return Err(Error::new_without_key(ErrorKind::MalformedValue {
-                            message: "code with scope length too small".to_string(),
-                        }));
+                        return Err(Error::new_without_key(ErrorKind::new_malformed(
+                            "code with scope length too small",
+                        )));
                     }
 
-                    self.verify_enough_bytes(valueoffset, length)?;
-                    let slice = &&self.doc.as_bytes()[valueoffset..(valueoffset + length)];
-                    let code = read_lenencoded(&slice[4..])?;
-                    let scope_start = 4 + 4 + code.len() + 1;
-                    let scope = RawDocument::from_bytes(&slice[scope_start..])?;
-                    (
-                        RawBsonRef::JavaScriptCodeWithScope(RawJavaScriptCodeWithScopeRef {
-                            code,
-                            scope,
-                        }),
-                        length,
-                    )
+                    length
                 }
                 ElementType::DbPointer => {
-                    let namespace = read_lenencoded(&self.doc.as_bytes()[valueoffset..])?;
-                    let id = self.next_oid(valueoffset + 4 + namespace.len() + 1)?;
-                    (
-                        RawBsonRef::DbPointer(RawDbPointerRef { namespace, id }),
-                        4 + namespace.len() + 1 + 12,
-                    )
+                    let length = read_len_and_end(&self.doc.as_bytes()[valueoffset..])?.0;
+                    length + 1 + 12
                 }
-                ElementType::Symbol => {
-                    let s = read_lenencoded(&self.doc.as_bytes()[valueoffset..])?;
-                    (RawBsonRef::Symbol(s), 4 + s.len() + 1)
-                }
-                ElementType::Decimal128 => {
-                    self.verify_enough_bytes(valueoffset, 16)?;
-                    (
-                        RawBsonRef::Decimal128(Decimal128::from_bytes(
-                            self.doc.as_bytes()[valueoffset..(valueoffset + 16)]
-                                .try_into()
-                                .unwrap(),
-                        )),
-                        16,
-                    )
-                }
-                ElementType::MinKey => (RawBsonRef::MinKey, 0),
-                ElementType::MaxKey => (RawBsonRef::MaxKey, 0),
+                ElementType::Decimal128 => 16,
+                ElementType::MinKey => 0,
+                ElementType::MaxKey => 0,
             };
 
-            self.offset = valueoffset + element_size;
             self.verify_enough_bytes(valueoffset, element_size)?;
+            self.offset = valueoffset + element_size;
 
-            Ok((key, element))
+            Ok((element_type, element_size))
         });
 
         if kvp_result.is_err() {
             self.valid = false;
         }
 
-        Some(kvp_result)
+        Some(match kvp_result {
+            Ok((kind, size)) => Ok(RawLazyElement {
+                key,
+                kind,
+                doc: self.doc,
+                start_at: valueoffset,
+                size: size,
+            }),
+            Err(e) => Err(e),
+        })
     }
 }
